Make sure entity exists before deleting
This caused an error when a workspace was added to detached object in
office form and then it was deleted without persisting the object. What
happened was that hibernate tried to delete an object that did not exist
in the database. This should now not happen for any entity in the app.

+++ src/main/java/com/bartoszthielmann/employeemanager/dao/employee/EmployeeDaoImpl.java
@@ -30,7 +30,9 @@ public class EmployeeDaoImpl implements EmployeeDao {
     @Override
     public void deleteById(int id) {
         Employee employee = entityManager.find(Employee.class, id);
-        entityManager.remove(employee);
+        if (employee != null) {
+            entityManager.remove(employee);
+        }
     }
 
     @Override
